fixed issue with player inventory
the prefab gameobject that corresponds with each stack of items in the player's inventory, was not being deleted upon using up the last item of a stack, only when dropping an item. added logic to the inventorymanager script's 'UseItem' method to include a step for deleting the gameobject upon final use.