implemented proper functionality for item placement
user key presses now cause placement of item, or escape from placement mode altogether. in the event the user escapes out of placement mode, the item is returned to the player's inventory. the item is removed from the datamanager inventory at placement only, so that in the event of a game crash during placement mode, no item is lost.