fixed issue with item placement
item placement was letting the player place items that were "inside" floors etc. fixed it to make sure that the length of the raycast hit was approximately equal to the distance to the ground for the item before marking it as placeable, ensuring that the item was properly being placed on the ground/floor and not within it.