fixed potential issue with second quest
if all of the dead trees had spawned as a single type of dead tree, the "sell 2 stacks of dead trees" would've been impossible to complete. Instead, the second quest is now updated for each dead tree sold, with a total required amount of 10. It now completes whether there are 2 stacks of dead trees, or 1.