updated quest manager with more quest reward logic
also fixed an issue causing the toggling of inventoryMenu from the questManager script to disallow any further toggling of the menu (the menu was being closed by the script before all of the drag and drop variables had a chance to be reset, causing the flags to fail upon trying to reopen it). Added the EndDrag() method to the ToggleInventoryMenu() method to fix this.