fixed issue with menus having camera movement when mouse moved
turned off mouse movement when inventory menu or trader menu is toggled on, allowing the user to freely use the mouse cursor without affecting the camera. Chose to use this solution instead of stopping time to allow background movement (from potential animals), as well as allowing timers and coroutines to keep running.