fixed issue with farming wood from trees
The wood was not correctly having its direction randomized as it was colliding with the tree object. Fixed this by creating an offset that the wood spawns at, in the direction of where it will ultimately be headed, to avoid colliding with the tree. Chose to use this method instead of fixing via ignoring the collision, as I wanted to leave the possibility of collisions with the tree at a time later than spawn (such as the wood bouncing off of the player and then hitting the tree again). Also added torque to the wood for a more fun effect.