fixed issue with item placement
item placement rotation now updates based on the position of the camera, to allow for more diversity in placed item rotation. Also fixed issue in item script which would cause a placeable item to be dropped into the world when picked up with a full inventory, rather than the intended action of being destroyed.