fixed issue with inventory system
the 'add additional item' method was incorrectly setting the quantity of an item in the data manager's list over the max stack quantity, while the item script itself prevented it from happening in-game. This caused an issue during scene transitions where the data manager was not able to correctly find an item, due to the quantities being inconsistent. Fixed the issue and added robustness to the method to prevent any further issues.