fixed issue where building was allowed when clipping through other buildable
changed the build manager script to disallow building if collision with a buildable object was detected, even if in an attachment slot. To solve the issue of attached builds naturally and intendedly colliding with each other, changed the collision check in the buildable object script to first register collisions, then if found, cast a overlap box check, with a radius slightly smaller than the actual object, to look for collisions that are more than just a little bit within the object. Box check only occurs when collisions are found in the collision array, to maintain efficiency.