fixed issue with footstep sounds
footsteps were playing even when the player was off the ground, so modified the player movement script to set footstep type to NoSound when IsGrounded == false. Also tweaked some other sounds and sound functionality, such as making it so the menu sounds only play when the user opens/closes the menu, but not when it is toggled by a script.